--- a/lib/cupti/CUPTI.jl
+++ b/lib/cupti/CUPTI.jl
@@ -5,7 +5,7 @@ using ..APIUtils
 using ..CUDA_Runtime
 
 using ..CUDA
-using ..CUDA: @retry_reclaim, initialize_context
+using ..CUDA: retry_reclaim, initialize_context
 using ..CUDA: CUuuid, CUcontext, CUstream, CUdevice, CUdevice_attribute,
               CUgraph, CUgraphNode, CUgraphNodeType, CUgraphExec, CUaccessPolicyWindow