--- a/lib/cusolver/libcusolver.jl
+++ b/lib/cusolver/libcusolver.jl
@@ -19,7 +19,9 @@ macro check(ex, errs...)
     end
 
     quote
-        res = @retry_reclaim err -> $check $(esc(ex))
+        res = retry_reclaim(err -> $check) do
+            $(esc(ex))
+        end
         if res != CUSOLVER_STATUS_SUCCESS
             throw_api_error(res)
         end